Change all integer types to Long and all floating point types to Double

+++ src/main/java/com/stripe/model/Coupon.java
@@ -20,14 +20,14 @@ public class Coupon extends ApiResource implements MetadataStore<Coupon>, HasId 
   Long created;
   String currency;
   String duration;
-  Integer durationInMonths;
+  Long durationInMonths;
   Boolean livemode;
   Long maxRedemptions;
   @Getter(onMethod = @__({@Override})) Map<String, String> metadata;
   String name;
-  Integer percentOff;
+  Long percentOff;
   Long redeemBy;
-  Integer timesRedeemed;
+  Long timesRedeemed;
   Boolean valid;
 
   // <editor-fold desc="create">
